requirements.txt
argparse
pyyaml
dnspython
requests
python-whois